+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminGoodsController.java
@@ -5,31 +5,15 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.linlinjava.litemall.admin.annotation.RequiresPermissionsDesc;
 import org.linlinjava.litemall.admin.dao.GoodsAllinone;
-import org.linlinjava.litemall.admin.util.CatVo;
-import org.linlinjava.litemall.core.qcode.QCodeService;
-import org.linlinjava.litemall.core.util.ResponseUtil;
+import org.linlinjava.litemall.admin.service.AdminGoodsService;
 import org.linlinjava.litemall.core.validator.Order;
 import org.linlinjava.litemall.core.validator.Sort;
-import org.linlinjava.litemall.db.domain.*;
-import org.linlinjava.litemall.db.service.*;
+import org.linlinjava.litemall.db.domain.LitemallGoods;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionDefinition;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
-import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import static org.linlinjava.litemall.admin.util.AdminResponseCode.GOODS_NAME_EXIST;
-import static org.linlinjava.litemall.admin.util.AdminResponseCode.GOODS_UPDATE_NOT_ALLOWED;
 
 @RestController
 @RequestMapping("/admin/goods")
@@ -38,366 +22,88 @@ public class AdminGoodsController {
     private final Log logger = LogFactory.getLog(AdminGoodsController.class);
 
     @Autowired
-    private PlatformTransactionManager txManager;
+    private AdminGoodsService adminGoodsService;
 
-    @Autowired
-    private LitemallGoodsService goodsService;
-    @Autowired
-    private LitemallGoodsSpecificationService specificationService;
-    @Autowired
-    private LitemallGoodsAttributeService attributeService;
-    @Autowired
-    private LitemallGoodsProductService productService;
-    @Autowired
-    private LitemallCategoryService categoryService;
-    @Autowired
-    private LitemallBrandService brandService;
-    @Autowired
-    private LitemallCartService cartService;
-    @Autowired
-    private LitemallOrderGoodsService orderGoodsService;
-
-    @Autowired
-    private QCodeService qCodeService;
-
+    /**
+     * 查询商品
+     *
+     * @param goodsSn
+     * @param name
+     * @param page
+     * @param limit
+     * @param sort
+     * @param order
+     * @return
+     */
     @RequiresPermissions("admin:goods:list")
-    @RequiresPermissionsDesc(menu={"商品管理" , "商品列表"}, button="查询")
+    @RequiresPermissionsDesc(menu = {"商品管理", "商品列表"}, button = "查询")
     @GetMapping("/list")
     public Object list(String goodsSn, String name,
                        @RequestParam(defaultValue = "1") Integer page,
                        @RequestParam(defaultValue = "10") Integer limit,
                        @Sort @RequestParam(defaultValue = "add_time") String sort,
                        @Order @RequestParam(defaultValue = "desc") String order) {
-        List<LitemallGoods> goodsList = goodsService.querySelective(goodsSn, name, page, limit, sort, order);
-        int total = goodsService.countSelective(goodsSn, name, page, limit, sort, order);
-        Map<String, Object> data = new HashMap<>();
-        data.put("total", total);
-        data.put("items", goodsList);
-
-        return ResponseUtil.ok(data);
-    }
-
-    private Object validate(GoodsAllinone goodsAllinone) {
-        LitemallGoods goods = goodsAllinone.getGoods();
-        String name = goods.getName();
-        if (StringUtils.isEmpty(name)) {
-            return ResponseUtil.badArgument();
-        }
-        String goodsSn = goods.getGoodsSn();
-        if (StringUtils.isEmpty(goodsSn)) {
-            return ResponseUtil.badArgument();
-        }
-        // 品牌商可以不设置，如果设置则需要验证品牌商存在
-        Integer brandId = goods.getBrandId();
-        if (brandId != null && brandId != 0) {
-            if (brandService.findById(brandId) == null) {
-                return ResponseUtil.badArgumentValue();
-            }
-        }
-        // 分类可以不设置，如果设置则需要验证分类存在
-        Integer categoryId = goods.getCategoryId();
-        if (categoryId != null && categoryId != 0) {
-            if (categoryService.findById(categoryId) == null) {
-                return ResponseUtil.badArgumentValue();
-            }
-        }
-
-        LitemallGoodsAttribute[] attributes = goodsAllinone.getAttributes();
-        for (LitemallGoodsAttribute attribute : attributes) {
-            String attr = attribute.getAttribute();
-            if (StringUtils.isEmpty(attr)) {
-                return ResponseUtil.badArgument();
-            }
-            String value = attribute.getValue();
-            if (StringUtils.isEmpty(value)) {
-                return ResponseUtil.badArgument();
-            }
-        }
-
-        LitemallGoodsSpecification[] specifications = goodsAllinone.getSpecifications();
-        for (LitemallGoodsSpecification specification : specifications) {
-            String spec = specification.getSpecification();
-            if (StringUtils.isEmpty(spec)) {
-                return ResponseUtil.badArgument();
-            }
-            String value = specification.getValue();
-            if (StringUtils.isEmpty(value)) {
-                return ResponseUtil.badArgument();
-            }
-        }
-
-        LitemallGoodsProduct[] products = goodsAllinone.getProducts();
-        for (LitemallGoodsProduct product : products) {
-            Integer number = product.getNumber();
-            if (number == null || number < 0) {
-                return ResponseUtil.badArgument();
-            }
-
-            BigDecimal price = product.getPrice();
-            if (price == null) {
-                return ResponseUtil.badArgument();
-            }
-
-            String[] productSpecifications = product.getSpecifications();
-            if (productSpecifications.length == 0) {
-                return ResponseUtil.badArgument();
-            }
-        }
-
-        return null;
+        return adminGoodsService.list(goodsSn, name, page, limit, sort, order);
     }
 
     /**
      * 编辑商品
-     * <p>
-     * TODO
-     * 目前商品修改的逻辑是
-     * 1. 更新litemall_goods表
-     * 2. 逻辑删除litemall_goods_specification、litemall_goods_attribute、litemall_goods_product
-     * 3. 添加litemall_goods_specification、litemall_goods_attribute、litemall_goods_product
      *
-     * 这里商品三个表的数据采用删除再添加的策略是因为
-     * 商品编辑页面，支持管理员添加删除商品规格、添加删除商品属性，因此这里仅仅更新是不可能的，
-     * 只能删除三个表旧的数据，然后添加新的数据。
-     * 但是这里又会引入新的问题，就是存在订单商品货品ID指向了失效的商品货品表。
-     * 因此这里会拒绝管理员编辑商品，如果订单或购物车中存在商品。
-     * 所以这里可能需要重新设计。
+     * @param goodsAllinone
+     * @return
      */
     @RequiresPermissions("admin:goods:update")
-    @RequiresPermissionsDesc(menu={"商品管理" , "商品列表"}, button="编辑")
+    @RequiresPermissionsDesc(menu = {"商品管理", "商品列表"}, button = "编辑")
     @PostMapping("/update")
     public Object update(@RequestBody GoodsAllinone goodsAllinone) {
-        Object error = validate(goodsAllinone);
-        if (error != null) {
-            return error;
-        }
-
-        LitemallGoods goods = goodsAllinone.getGoods();
-        LitemallGoodsAttribute[] attributes = goodsAllinone.getAttributes();
-        LitemallGoodsSpecification[] specifications = goodsAllinone.getSpecifications();
-        LitemallGoodsProduct[] products = goodsAllinone.getProducts();
-
-        Integer id = goods.getId();
-        // 检查是否存在购物车商品或者订单商品
-        // 如果存在则拒绝修改商品。
-        if(orderGoodsService.checkExist(id)){
-            return ResponseUtil.fail(GOODS_UPDATE_NOT_ALLOWED, "商品已经在订单中，不能修改");
-        }
-        if(cartService.checkExist(id)){
-            return ResponseUtil.fail(GOODS_UPDATE_NOT_ALLOWED, "商品已经在购物车中，不能修改");
-        }
-
-        // 开启事务管理
-        DefaultTransactionDefinition def = new DefaultTransactionDefinition();
-        def.setPropagationBehavior(TransactionDefinition.PROPAGATION_REQUIRED);
-        TransactionStatus status = txManager.getTransaction(def);
-        try {
-
-            //将生成的分享图片地址写入数据库
-            String url = qCodeService.createGoodShareImage(goods.getId().toString(), goods.getPicUrl(), goods.getName());
-            goods.setShareUrl(url);
-
-            // 商品基本信息表litemall_goods
-            if (goodsService.updateById(goods) == 0) {
-                throw new Exception("更新数据失败");
-            }
-
-            Integer gid = goods.getId();
-            specificationService.deleteByGid(gid);
-            attributeService.deleteByGid(gid);
-            productService.deleteByGid(gid);
-
-            // 商品规格表litemall_goods_specification
-            for (LitemallGoodsSpecification specification : specifications) {
-                specification.setGoodsId(goods.getId());
-                specificationService.add(specification);
-            }
-
-            // 商品参数表litemall_goods_attribute
-            for (LitemallGoodsAttribute attribute : attributes) {
-                attribute.setGoodsId(goods.getId());
-                attributeService.add(attribute);
-            }
-
-            // 商品货品表litemall_product
-            for (LitemallGoodsProduct product : products) {
-                product.setGoodsId(goods.getId());
-                productService.add(product);
-            }
-        } catch (Exception ex) {
-            txManager.rollback(status);
-            logger.error("系统内部错误", ex);
-            return ResponseUtil.fail();
-        }
-        txManager.commit(status);
-
-        qCodeService.createGoodShareImage(goods.getId().toString(), goods.getPicUrl(), goods.getName());
-
-        return ResponseUtil.ok();
+        return adminGoodsService.update(goodsAllinone);
     }
 
+    /**
+     * 删除商品
+     *
+     * @param goods
+     * @return
+     */
     @RequiresPermissions("admin:goods:delete")
-    @RequiresPermissionsDesc(menu={"商品管理" , "商品列表"}, button="删除")
+    @RequiresPermissionsDesc(menu = {"商品管理", "商品列表"}, button = "删除")
     @PostMapping("/delete")
     public Object delete(@RequestBody LitemallGoods goods) {
-        Integer id = goods.getId();
-        if (id == null) {
-            return ResponseUtil.badArgument();
-        }
-
-        // 开启事务管理
-        DefaultTransactionDefinition def = new DefaultTransactionDefinition();
-        def.setPropagationBehavior(TransactionDefinition.PROPAGATION_REQUIRED);
-        TransactionStatus status = txManager.getTransaction(def);
-        try {
-
-            Integer gid = goods.getId();
-            goodsService.deleteById(gid);
-            specificationService.deleteByGid(gid);
-            attributeService.deleteByGid(gid);
-            productService.deleteByGid(gid);
-        } catch (Exception ex) {
-            txManager.rollback(status);
-            logger.error("系统内部错误", ex);
-            return ResponseUtil.fail();
-        }
-        txManager.commit(status);
-        return ResponseUtil.ok();
+        return adminGoodsService.delete(goods);
     }
 
+    /**
+     * 添加商品
+     *
+     * @param goodsAllinone
+     * @return
+     */
     @RequiresPermissions("admin:goods:create")
-    @RequiresPermissionsDesc(menu={"商品管理" , "商品上架"}, button="上架")
+    @RequiresPermissionsDesc(menu = {"商品管理", "商品上架"}, button = "上架")
     @PostMapping("/create")
     public Object create(@RequestBody GoodsAllinone goodsAllinone) {
-        Object error = validate(goodsAllinone);
-        if (error != null) {
-            return error;
-        }
-
-        LitemallGoods goods = goodsAllinone.getGoods();
-        LitemallGoodsAttribute[] attributes = goodsAllinone.getAttributes();
-        LitemallGoodsSpecification[] specifications = goodsAllinone.getSpecifications();
-        LitemallGoodsProduct[] products = goodsAllinone.getProducts();
-
-        String name = goods.getName();
-        if (goodsService.checkExistByName(name)) {
-            return ResponseUtil.fail(GOODS_NAME_EXIST, "商品名已经存在");
-        }
-
-        // 开启事务管理
-        DefaultTransactionDefinition def = new DefaultTransactionDefinition();
-        def.setPropagationBehavior(TransactionDefinition.PROPAGATION_REQUIRED);
-        TransactionStatus status = txManager.getTransaction(def);
-        try {
-
-            // 商品基本信息表litemall_goods
-            goodsService.add(goods);
-
-            //将生成的分享图片地址写入数据库
-            String url = qCodeService.createGoodShareImage(goods.getId().toString(), goods.getPicUrl(), goods.getName());
-            if (!StringUtils.isEmpty(url)) {
-                goods.setShareUrl(url);
-                if (goodsService.updateById(goods) == 0) {
-                    throw new Exception("更新数据失败");
-                }
-            }
-
-            // 商品规格表litemall_goods_specification
-            for (LitemallGoodsSpecification specification : specifications) {
-                specification.setGoodsId(goods.getId());
-                specificationService.add(specification);
-            }
-
-            // 商品参数表litemall_goods_attribute
-            for (LitemallGoodsAttribute attribute : attributes) {
-                attribute.setGoodsId(goods.getId());
-                attributeService.add(attribute);
-            }
-
-            // 商品货品表litemall_product
-            for (LitemallGoodsProduct product : products) {
-                product.setGoodsId(goods.getId());
-                productService.add(product);
-            }
-        } catch (Exception ex) {
-            txManager.rollback(status);
-            logger.error("系统内部错误", ex);
-            return ResponseUtil.fail();
-        }
-        txManager.commit(status);
-
-        return ResponseUtil.ok();
+        return adminGoodsService.create(goodsAllinone);
     }
 
     @RequiresPermissions("admin:goods:list")
-    @RequiresPermissionsDesc(menu={"商品管理" , "商品列表"}, button="查询")
+    @RequiresPermissionsDesc(menu = {"商品管理", "商品列表"}, button = "查询")
     @GetMapping("/catAndBrand")
     public Object list2() {
-        // http://element-cn.eleme.io/#/zh-CN/component/cascader
-        // 管理员设置“所属分类”
-        List<LitemallCategory> l1CatList = categoryService.queryL1();
-        List<CatVo> categoryList = new ArrayList<>(l1CatList.size());
-
-        for (LitemallCategory l1 : l1CatList) {
-            CatVo l1CatVo = new CatVo();
-            l1CatVo.setValue(l1.getId());
-            l1CatVo.setLabel(l1.getName());
-
-            List<LitemallCategory> l2CatList = categoryService.queryByPid(l1.getId());
-            List<CatVo> children = new ArrayList<>(l2CatList.size());
-            for (LitemallCategory l2 : l2CatList) {
-                CatVo l2CatVo = new CatVo();
-                l2CatVo.setValue(l2.getId());
-                l2CatVo.setLabel(l2.getName());
-                children.add(l2CatVo);
-            }
-            l1CatVo.setChildren(children);
-
-            categoryList.add(l1CatVo);
-        }
-
-        // http://element-cn.eleme.io/#/zh-CN/component/select
-        // 管理员设置“所属品牌商”
-        List<LitemallBrand> list = brandService.all();
-        List<Map<String, Object>> brandList = new ArrayList<>(l1CatList.size());
-        for (LitemallBrand brand : list) {
-            Map<String, Object> b = new HashMap<>(2);
-            b.put("value", brand.getId());
-            b.put("label", brand.getName());
-            brandList.add(b);
-        }
-
-        Map<String, Object> data = new HashMap<>();
-        data.put("categoryList", categoryList);
-        data.put("brandList", brandList);
-        return ResponseUtil.ok(data);
+        return adminGoodsService.list2();
     }
 
+    /**
+     * 商品详情
+     *
+     * @param id
+     * @return
+     */
     @RequiresPermissions("admin:goods:read")
-    @RequiresPermissionsDesc(menu={"商品管理" , "商品列表"}, button="编辑")
+    @RequiresPermissionsDesc(menu = {"商品管理", "商品列表"}, button = "编辑")
     @GetMapping("/detail")
     public Object detail(@NotNull Integer id) {
-        LitemallGoods goods = goodsService.findById(id);
-        List<LitemallGoodsProduct> products = productService.queryByGid(id);
-        List<LitemallGoodsSpecification> specifications = specificationService.queryByGid(id);
-        List<LitemallGoodsAttribute> attributes = attributeService.queryByGid(id);
+        return adminGoodsService.detail(id);
 
-        Integer categoryId = goods.getCategoryId();
-        LitemallCategory category = categoryService.findById(categoryId);
-        Integer[] categoryIds = new Integer[]{};
-        if (category != null) {
-            Integer parentCategoryId = category.getPid();
-            categoryIds = new Integer[]{parentCategoryId, categoryId};
-        }
-
-        Map<String, Object> data = new HashMap<>();
-        data.put("goods", goods);
-        data.put("specifications", specifications);
-        data.put("products", products);
-        data.put("attributes", attributes);
-        data.put("categoryIds", categoryIds);
-
-        return ResponseUtil.ok(data);
     }
 
 }
