+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminOrderController.java
@@ -1,42 +1,18 @@
 package org.linlinjava.litemall.admin.web;
 
-import com.github.binarywang.wxpay.bean.request.WxPayRefundRequest;
-import com.github.binarywang.wxpay.bean.result.WxPayRefundResult;
-import com.github.binarywang.wxpay.exception.WxPayException;
-import com.github.binarywang.wxpay.service.WxPayService;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.linlinjava.litemall.admin.annotation.RequiresPermissionsDesc;
-import org.linlinjava.litemall.core.notify.NotifyService;
-import org.linlinjava.litemall.core.notify.NotifyType;
-import org.linlinjava.litemall.core.util.JacksonUtil;
-import org.linlinjava.litemall.core.util.ResponseUtil;
+import org.linlinjava.litemall.admin.service.AdminOrderService;
 import org.linlinjava.litemall.core.validator.Order;
 import org.linlinjava.litemall.core.validator.Sort;
-import org.linlinjava.litemall.db.domain.LitemallComment;
-import org.linlinjava.litemall.db.domain.LitemallOrder;
-import org.linlinjava.litemall.db.domain.LitemallOrderGoods;
-import org.linlinjava.litemall.db.domain.UserVo;
-import org.linlinjava.litemall.db.service.*;
-import org.linlinjava.litemall.db.util.OrderUtil;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionDefinition;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
-import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.constraints.NotNull;
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-
-import static org.linlinjava.litemall.admin.util.AdminResponseCode.*;
 
 @RestController
 @RequestMapping("/admin/order")
@@ -45,25 +21,22 @@ public class AdminOrderController {
     private final Log logger = LogFactory.getLog(AdminOrderController.class);
 
     @Autowired
-    private PlatformTransactionManager txManager;
+    private AdminOrderService adminOrderService;
 
-    @Autowired
-    private LitemallOrderGoodsService orderGoodsService;
-    @Autowired
-    private LitemallOrderService orderService;
-    @Autowired
-    private LitemallGoodsProductService productService;
-    @Autowired
-    private LitemallUserService userService;
-    @Autowired
-    private LitemallCommentService commentService;
-    @Autowired
-    private WxPayService wxPayService;
-    @Autowired
-    private NotifyService notifyService;
-
+    /**
+     * 查询订单
+     *
+     * @param userId
+     * @param orderSn
+     * @param orderStatusArray
+     * @param page
+     * @param limit
+     * @param sort
+     * @param order
+     * @return
+     */
     @RequiresPermissions("admin:order:list")
-    @RequiresPermissionsDesc(menu={"商城管理" , "订单管理"}, button="查询")
+    @RequiresPermissionsDesc(menu = {"商城管理", "订单管理"}, button = "查询")
     @GetMapping("/list")
     public Object list(Integer userId, String orderSn,
                        @RequestParam(required = false) List<Short> orderStatusArray,
@@ -71,218 +44,60 @@ public class AdminOrderController {
                        @RequestParam(defaultValue = "10") Integer limit,
                        @Sort @RequestParam(defaultValue = "add_time") String sort,
                        @Order @RequestParam(defaultValue = "desc") String order) {
-        List<LitemallOrder> orderList = orderService.querySelective(userId, orderSn, orderStatusArray, page, limit, sort, order);
-        int total = orderService.countSelective(userId, orderSn, orderStatusArray, page, limit, sort, order);
-
-        Map<String, Object> data = new HashMap<>();
-        data.put("total", total);
-        data.put("items", orderList);
-
-        return ResponseUtil.ok(data);
+        return adminOrderService.list(userId, orderSn, orderStatusArray, page, limit, sort, order);
     }
 
+    /**
+     * 订单详情
+     *
+     * @param id
+     * @return
+     */
     @RequiresPermissions("admin:order:read")
-    @RequiresPermissionsDesc(menu={"商城管理" , "订单管理"}, button="详情")
+    @RequiresPermissionsDesc(menu = {"商城管理", "订单管理"}, button = "详情")
     @GetMapping("/detail")
     public Object detail(@NotNull Integer id) {
-        LitemallOrder order = orderService.findById(id);
-        List<LitemallOrderGoods> orderGoods = orderGoodsService.queryByOid(id);
-        UserVo user = userService.findUserVoById(order.getUserId());
-        Map<String, Object> data = new HashMap<>();
-        data.put("order", order);
-        data.put("orderGoods", orderGoods);
-        data.put("user", user);
-
-        return ResponseUtil.ok(data);
+        return adminOrderService.detail(id);
     }
 
     /**
      * 订单退款
-     * <p>
-     * 1. 检测当前订单是否能够退款;
-     * 2. 微信退款操作;
-     * 3. 设置订单退款确认状态；
-     * 4. 订单商品库存回库。
-     * <p>
-     * TODO
-     * 虽然接入了微信退款API，但是从安全角度考虑，建议开发者删除这里微信退款代码，采用以下两步走步骤：
-     * 1. 管理员登录微信官方支付平台点击退款操作进行退款
-     * 2. 管理员登录litemall管理后台点击退款操作进行订单状态修改和商品库存回库
      *
-     * @param body    订单信息，{ orderId：xxx }
+     * @param body 订单信息，{ orderId：xxx }
      * @return 订单退款操作结果
      */
     @RequiresPermissions("admin:order:refund")
-    @RequiresPermissionsDesc(menu={"商城管理" , "订单管理"}, button="订单退款")
+    @RequiresPermissionsDesc(menu = {"商城管理", "订单管理"}, button = "订单退款")
     @PostMapping("refund")
     public Object refund(@RequestBody String body) {
-        Integer orderId = JacksonUtil.parseInteger(body, "orderId");
-        String refundMoney = JacksonUtil.parseString(body, "refundMoney");
-        if (orderId == null) {
-            return ResponseUtil.badArgument();
-        }
-        if(StringUtils.isEmpty(refundMoney)){
-            return ResponseUtil.badArgument();
-        }
-
-        LitemallOrder order = orderService.findById(orderId);
-        if (order == null) {
-            return ResponseUtil.badArgument();
-        }
-
-        if (order.getActualPrice().compareTo(new BigDecimal(refundMoney)) != 0) {
-            return ResponseUtil.badArgumentValue();
-        }
-
-        // 如果订单不是退款状态，则不能退款
-        if (!order.getOrderStatus().equals(OrderUtil.STATUS_REFUND)) {
-            return ResponseUtil.fail(ORDER_CONFIRM_NOT_ALLOWED, "订单不能确认收货");
-        }
-
-        // 微信退款
-        WxPayRefundRequest wxPayRefundRequest = new WxPayRefundRequest();
-        wxPayRefundRequest.setOutTradeNo(order.getOrderSn());
-        wxPayRefundRequest.setOutRefundNo("refund_" + order.getOrderSn());
-        // 元转成分
-        Integer totalFee = order.getActualPrice().multiply(new BigDecimal(100)).intValue();
-        wxPayRefundRequest.setTotalFee(totalFee);
-        wxPayRefundRequest.setRefundFee(totalFee);
-
-        WxPayRefundResult wxPayRefundResult = null;
-        try {
-            wxPayRefundResult = wxPayService.refund(wxPayRefundRequest);
-        } catch (WxPayException e) {
-            e.printStackTrace();
-            return ResponseUtil.fail(ORDER_REFUND_FAILED, "订单退款失败");
-        }
-        if(!wxPayRefundResult.getReturnCode().equals("SUCCESS")){
-            logger.warn("refund fail: " +  wxPayRefundResult.getReturnMsg());
-            return ResponseUtil.fail(ORDER_REFUND_FAILED, "订单退款失败");
-        }
-        if(!wxPayRefundResult.getResultCode().equals("SUCCESS")){
-            logger.warn("refund fail: " +  wxPayRefundResult.getReturnMsg());
-            return ResponseUtil.fail(ORDER_REFUND_FAILED, "订单退款失败");
-        }
-
-        // 开启事务管理
-        DefaultTransactionDefinition def = new DefaultTransactionDefinition();
-        def.setPropagationBehavior(TransactionDefinition.PROPAGATION_REQUIRED);
-        TransactionStatus status = txManager.getTransaction(def);
-        try {
-            // 设置订单取消状态
-            order.setOrderStatus(OrderUtil.STATUS_REFUND_CONFIRM);
-            if (orderService.updateWithOptimisticLocker(order) == 0) {
-                throw new Exception("更新数据已失效");
-            }
-
-            // 商品货品数量增加
-            List<LitemallOrderGoods> orderGoodsList = orderGoodsService.queryByOid(orderId);
-            for (LitemallOrderGoods orderGoods : orderGoodsList) {
-                Integer productId = orderGoods.getProductId();
-                Short number = orderGoods.getNumber();
-                if (productService.addStock(productId, number) == 0) {
-                    throw new Exception("商品货品库存增加失败");
-                }
-            }
-        } catch (Exception ex) {
-            txManager.rollback(status);
-            logger.error("系统内部错误", ex);
-            return ResponseUtil.fail(ORDER_REFUND_FAILED, "订单退款失败");
-        }
-        txManager.commit(status);
-
-        //TODO 发送邮件和短信通知，这里采用异步发送
-        // 退款成功通知用户, 例如“您申请的订单退款 [ 单号:{1} ] 已成功，请耐心等待到账。”
-        // 注意订单号只发后6位
-        notifyService.notifySmsTemplate(order.getMobile(), NotifyType.REFUND, new String[]{order.getOrderSn().substring(8, 14)});
-
-        return ResponseUtil.ok();
+        return adminOrderService.refund(body);
     }
 
     /**
      * 发货
-     * 1. 检测当前订单是否能够发货
-     * 2. 设置订单发货状态
      *
-     * @param body    订单信息，{ orderId：xxx, shipSn: xxx, shipChannel: xxx }
+     * @param body 订单信息，{ orderId：xxx, shipSn: xxx, shipChannel: xxx }
      * @return 订单操作结果
-     * 成功则 { errno: 0, errmsg: '成功' }
-     * 失败则 { errno: XXX, errmsg: XXX }
      */
     @RequiresPermissions("admin:order:ship")
-    @RequiresPermissionsDesc(menu={"商城管理" , "订单管理"}, button="订单发货")
+    @RequiresPermissionsDesc(menu = {"商城管理", "订单管理"}, button = "订单发货")
     @PostMapping("ship")
     public Object ship(@RequestBody String body) {
-        Integer orderId = JacksonUtil.parseInteger(body, "orderId");
-        String shipSn = JacksonUtil.parseString(body, "shipSn");
-        String shipChannel = JacksonUtil.parseString(body, "shipChannel");
-        if (orderId == null || shipSn == null || shipChannel == null) {
-            return ResponseUtil.badArgument();
-        }
-
-        LitemallOrder order = orderService.findById(orderId);
-        if (order == null) {
-            return ResponseUtil.badArgument();
-        }
-
-        // 如果订单不是已付款状态，则不能发货
-        if (!order.getOrderStatus().equals(OrderUtil.STATUS_PAY)) {
-            return ResponseUtil.fail(ORDER_CONFIRM_NOT_ALLOWED, "订单不能确认收货");
-        }
-
-        order.setOrderStatus(OrderUtil.STATUS_SHIP);
-        order.setShipSn(shipSn);
-        order.setShipChannel(shipChannel);
-        order.setShipTime(LocalDateTime.now());
-        if (orderService.updateWithOptimisticLocker(order) == 0) {
-            return ResponseUtil.updatedDateExpired();
-        }
-
-        //TODO 发送邮件和短信通知，这里采用异步发送
-        // 发货会发送通知短信给用户:          *
-        // "您的订单已经发货，快递公司 {1}，快递单 {2} ，请注意查收"
-        notifyService.notifySmsTemplate(order.getMobile(), NotifyType.SHIP, new String[]{shipChannel, shipSn});
-
-        return ResponseUtil.ok();
+        return adminOrderService.ship(body);
     }
 
 
     /**
      * 回复订单商品
      *
-     * @param body    订单信息，{ orderId：xxx }
+     * @param body 订单信息，{ orderId：xxx }
      * @return 订单操作结果
-     * 成功则 { errno: 0, errmsg: '成功' }
-     * 失败则 { errno: XXX, errmsg: XXX }
      */
     @RequiresPermissions("admin:order:reply")
-    @RequiresPermissionsDesc(menu={"商城管理" , "订单管理"}, button="订单商品回复")
+    @RequiresPermissionsDesc(menu = {"商城管理", "订单管理"}, button = "订单商品回复")
     @PostMapping("reply")
     public Object reply(@RequestBody String body) {
-        Integer commentId = JacksonUtil.parseInteger(body, "commentId");
-        if (commentId == null || commentId == 0) {
-            return ResponseUtil.badArgument();
-        }
-        // 目前只支持回复一次
-        if (commentService.findById(commentId) != null) {
-            return ResponseUtil.fail(ORDER_REPLY_EXIST, "订单商品已回复！");
-        }
-        String content = JacksonUtil.parseString(body, "content");
-        if (StringUtils.isEmpty(content)) {
-            return ResponseUtil.badArgument();
-        }
-        // 创建评价回复
-        LitemallComment comment = new LitemallComment();
-        comment.setType((byte) 2);
-        comment.setValueId(commentId);
-        comment.setContent(content);
-        comment.setUserId(0);                 // 评价回复没有用
-        comment.setStar((short) 0);           // 评价回复没有用
-        comment.setHasPicture(false);        // 评价回复没有用
-        comment.setPicUrls(new String[]{});  // 评价回复没有用
-        commentService.save(comment);
-
-        return ResponseUtil.ok();
+        return adminOrderService.reply(body);
     }
 
 }
